feat: add toast alerts
