+++ frontend/package.json
@@ -19,6 +19,7 @@
     "react-dom": "^19.1.0",
     "react-i18next": "^15.7.0",
     "react-router-dom": "^7.8.1",
+    "react-toastify": "^11.0.5",
     "recharts": "^3.1.2"
   },
   "devDependencies": {
