+++ frontend/src/pages/Alerts.tsx
@@ -1,0 +1,76 @@
+import * as api from "../api";
+import type { Alert } from "../types";
+import { useFetch } from "../hooks/useFetch";
+import { useEffect, useState } from "react";
+
+export default function Alerts({ user = "default" }: { user?: string }) {
+  const { data: alerts, loading, error } = useFetch<Alert[]>(api.getAlerts, []);
+  const [threshold, setThreshold] = useState<number>();
+  const [settingsError, setSettingsError] = useState(false);
+
+  useEffect(() => {
+    if (api.getAlertSettings) {
+      api
+        .getAlertSettings(user)
+        .then((r) => setThreshold(r.threshold))
+        .catch(() => setSettingsError(true));
+    }
+  }, [user]);
+
+  const save = () => {
+    if (threshold !== undefined && api.setAlertSettings) {
+      api.setAlertSettings(user, threshold);
+    }
+  };
+
+  if (loading) return null;
+
+  if (error || settingsError) {
+    return (
+      <div style={{ border: "1px solid #ccc", padding: "0.5rem", marginBottom: "1rem" }}>
+        Cannot reach server
+      </div>
+    );
+  }
+
+  const importAlert = alerts?.find((a) => a.ticker === "IMPORT");
+  const otherAlerts = alerts?.filter((a) => a.ticker !== "IMPORT") ?? [];
+
+  if (!importAlert && otherAlerts.length === 0) return <div>No alerts</div>;
+
+  return (
+    <div style={{ maxWidth: 900, margin: "0 auto", padding: "1rem" }}>
+      <h1>Alerts</h1>
+      {importAlert && (
+        <div data-testid="import-status" style={{ marginTop: "0.5rem" }}>
+          {importAlert.message}
+        </div>
+      )}
+      {otherAlerts.length > 0 && (
+        <>
+          <div style={{ marginTop: "0.5rem" }}>
+            <label>
+              Threshold %:{" "}
+              <input
+                type="number"
+                value={threshold ?? ""}
+                onChange={(e) => setThreshold(parseFloat(e.target.value))}
+                style={{ width: "4rem" }}
+              />
+            </label>
+            <button onClick={save} style={{ marginLeft: "0.5rem" }}>
+              Save
+            </button>
+          </div>
+          <ul style={{ margin: 0, paddingLeft: "1.2rem" }}>
+            {otherAlerts.map((a, i) => (
+              <li key={i}>
+                <strong>{a.ticker}</strong>: {a.message}
+              </li>
+            ))}
+          </ul>
+        </>
+      )}
+    </div>
+  );
+}
